pyproject.toml
[tool.poetry]
name = "django-twined"
version = "0.3.0"
description = "A django app to manage octue services"
authors = ["Tom Clark <tom@octue.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
repository = "https://github.com/octue/django-twined"
keywords = ["django", "services", "octue", "twined"]
packages = [{ include = "django_twined" },]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
django = ">=3.0,<4.0"
channels = ">=3.0,<4.0"
octue = "^0.39.0"
pika = "^1.2.0"
django-gcp = "^0.7.3"
django-jsoneditor = "^0.2.2"
django-model-utils = "^4.2.0"
django-modelclone = "^0.7.1"

[tool.poetry.dev-dependencies]
channels = ">=3.0,<=4.0"
coverage = "^6.2"
pre-commit = "^2.17.0"
pytest-sugar = "^0.9.4"
pytest-django = "^4.5.2"
pytest-asyncio = "^0.18.1"
pytest = "^7.0.0"
tox = "^3.24.5"
tox-gh-actions = "^2.9.1"
tox-poetry = "^0.4.1"
sphinx-rtd-theme = "^1.0.0"
sphinx-tabs = "^3.2.0"
sphinx-charts = "^0.1.2"
Sphinx = "^4.4.0"
django-dramatiq = "^0.10.0"
psycopg2 = "^2.9.3"
django-test-migrations = "^1.2.0"
pylint-django = "^2.5.3"
django-extensions = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
